fix bug

+++ src/main/scala/com/cpc/spark/log/report/GetHourReport.scala
@@ -581,14 +581,14 @@ object GetHourReport {
           var isclick = x.getAs[Int]("isclick")
           var spam_click = x.getAs[Int]("spam_click")
           var antispam_score = x.getAs[Int]("antispam_score")
-          var realCost = 0L
+          var realCost = 0
           if (isclick > 0 && antispam_score == 10000) {
-            realCost = x.getAs[Long]("price")
+            realCost = x.getAs[Int]("price")
             if (realCost >= Int.MaxValue) {
               realCost = 1
             }
           } else {
-            realCost = 0L
+            realCost = 0
           }
           val charge = MediaChargeReport( //adslotType = x.getAs[Int]("adslot_type")
             //media_id = x.getAs[String]("media_appsid").toInt,
@@ -605,7 +605,7 @@ object GetHourReport {
             click = isclick,
             //charged_click = isclick,
             //spam_click = spam_click,
-            cash_cost = realCost.toInt,
+            cash_cost = realCost,
             date = x.getAs[String]("date"),
             hour = x.getAs[String]("hour").toInt
           )
